# specs/001-local-prompt-manager/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Prompt Management Local API
  version: 1.0.0
  description: 本機提示詞管理 Route Handlers 合約（Next.js REST）
servers:
  - url: http://localhost:3000/api
paths:
  /projects:
    get:
      summary: 列出專案
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/ProjectStatusCode" }
        - in: query
          name: projectType
          schema: { type: string, description: "taxonomy code" }
        - in: query
          name: tag
          schema: { type: string, description: "taxonomy code" }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      summary: 新增專案
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: 名稱重複
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - type: object
                    properties:
                      details:
                        $ref: "#/components/schemas/ConflictFieldDetails"
    patch:
      summary: 更新專案（名稱/狀態/提示詞數量）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: 名稱重複
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - type: object
                    properties:
                      details:
                        $ref: "#/components/schemas/ConflictFieldDetails"
    delete:
      summary: 刪除專案
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /prompts:
    get:
      summary: 列出提示詞（支援篩選/搜尋）
      parameters:
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string, description: "taxonomy code" }
        - in: query
          name: promptStage
          schema: { type: string, description: "taxonomy code" }
        - in: query
          name: platformTag
          schema: { type: string, description: "taxonomy code" }
        - in: query
          name: tag
          schema: { type: string, description: "taxonomy code" }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptListItem"
    post:
      summary: 新增提示詞
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptDetail"
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /inbox:
    get:
      summary: 列出收件匣草稿（可分頁/搜尋）
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxList"
    post:
      summary: 新增收件匣草稿
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
...
components:
...
  responses:
    BadRequest:
      description: 驗證失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    TaxonomyValue:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
    ProjectStatusCode:
      type: string
      enum: [ACTIVE, PAUSED, ARCHIVED]
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: e.g. bad_request | not_found | conflict }
        message: { type: string }
        details: { description: 錯誤細節（依端點而異） }
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    IsoDateTime:
      type: string
      format: date-time
      description: ISO 8601（UTC+08:00）
    ConflictFieldDetails:
      type: object
      properties:
        field: { type: string }
...
    Project:
      type: object
      required:
        [id, name, status, summary, projectType, tags, promptCount, docPath, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        status: { $ref: "#/components/schemas/TaxonomyValue" }
        summary: { type: string }
        projectType: { $ref: "#/components/schemas/TaxonomyValue" }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        promptCount: { type: integer }
        docPath: { type: string }
        path: { type: string }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        updatedAt: { $ref: "#/components/schemas/IsoDateTime" }
    ProjectCreateInput:
      type: object
      required: [name, status, summary, projectType, tags]
      properties:
        id: { type: string }
        name: { type: string }
        status: { $ref: "#/components/schemas/TaxonomyValue" }
        summary: { type: string }
        projectType: { $ref: "#/components/schemas/TaxonomyValue" }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
    ProjectUpdateInput:
      type: object
      required: [id]
      properties:
        id: { type: string }
        name: { type: string }
        status: { $ref: "#/components/schemas/TaxonomyValue" }
        summary: { type: string }
        projectType: { $ref: "#/components/schemas/TaxonomyValue" }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        promptCount: { type: integer }
...
    PromptFrontmatter:
      type: object
      required:
        [title, project, type, status, category, promptStage, platformTags, audienceTags, deliverableTags, tags, createdAt, updatedAt]
      properties:
        title: { type: string }
        project: { type: string }
        type: { type: string }
        status: { type: string }
        category: { $ref: "#/components/schemas/TaxonomyValue" }
        promptStage: { $ref: "#/components/schemas/TaxonomyValue" }
        model: { type: string }
        platformTags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        audienceTags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        deliverableTags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        note: { type: string }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        updatedAt: { $ref: "#/components/schemas/IsoDateTime" }
    PromptListItem:
      type: object
      required: [id, projectId, title, type, status, tags, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        path: { type: string }
        title: { type: string }
        project: { type: string }
        type: { type: string }
        status: { type: string }
        model: { type: string }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyValue" }
        note: { type: string }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        updatedAt: { $ref: "#/components/schemas/IsoDateTime" }
        preview: { type: string }
    PromptDetail:
      type: object
      required: [id, projectId, frontmatter, body, hash, mtimeMs]
      properties:
        id: { type: string }
        projectId: { type: string }
        frontmatter: { $ref: "#/components/schemas/PromptFrontmatter" }
        body: { type: string }
        hash: { type: string }
        mtimeMs: { type: number }
    PromptCreateInput:
      type: object
      required: [frontmatter]
      properties:
        frontmatter: { $ref: "#/components/schemas/PromptFrontmatter" }
        body: { type: string }
...
    InboxItem:
      type: object
      required: [id, title, content, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        content: { type: string }
        hint: { type: string }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        updatedAt: { $ref: "#/components/schemas/IsoDateTime" }
    InboxList:
      type: object
      required: [items, total, hasMore, limit, offset]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InboxItem" }
        total: { type: integer }
        hasMore: { type: boolean }
        limit: { type: integer }
        offset: { type: integer }
    InboxCreateInput:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        hint: { type: string }
...